feat: Add timestamp argument to STEP export
Allow passing a timestamp value when export STEP files, to generate STEP
files with a specific timestamp value in the file header.

For example, a null timestamp (`0000-00-00T00:00:00`), or a static
timestamp can be used when generated files should be equal if there are
no visual changes, such as for file versioning.

This commit extends the `#export_step` function, to accept a `timestamp`
keyword argument, that can be a string or a `datetime` object. A
`datetime` is easier to use from Python.